+++ packages/stream_chat_flutter/example/lib/tutorial_part_3.dart
@@ -146,6 +146,7 @@ class _ChannelListPageState extends State<ChannelListPage> {
               color: StreamChatTheme.of(context)
                   .colorTheme
                   .textHighEmphasis
+                  // ignore: deprecated_member_use
                   .withOpacity(opacity),
             ),
         channel: channel,