+++ packages/stream_chat_flutter/lib/src/message_input/stream_message_input_attachment_list.dart
@@ -357,6 +357,7 @@ class RemoveAttachmentButton extends StatelessWidget {
         shape: RoundedRectangleBorder(
           borderRadius: BorderRadius.circular(16),
         ),
+        // ignore: deprecated_member_use
         fillColor: colorTheme.textHighEmphasis.withOpacity(0.5),
         child: StreamSvgIcon.close(
           size: 24,