+++ packages/stream_chat_flutter/lib/src/message_widget/giphy_ephemeral_message.dart
@@ -222,6 +222,7 @@ class GiphyHeader extends StatelessWidget {
               maxLines: 1,
               overflow: TextOverflow.ellipsis,
               style: TextStyle(
+                // ignore: deprecated_member_use
                 color: colorTheme.textHighEmphasis.withOpacity(0.5),
               ),
             ),