+++ packages/stream_chat_flutter/lib/src/scroll_view/message_search_scroll_view/stream_message_search_list_view.dart
@@ -378,6 +378,7 @@ class StreamMessageSearchListSeparator extends StatelessWidget {
     final effect = StreamChatTheme.of(context).colorTheme.borderBottom;
     return Container(
       height: 1,
+      // ignore: deprecated_member_use
       color: effect.color!.withOpacity(effect.alpha ?? 1.0),
     );
   }