+++ packages/stream_chat_flutter/lib/src/scroll_view/user_scroll_view/stream_user_list_view.dart
@@ -372,6 +372,7 @@ class StreamUserListSeparator extends StatelessWidget {
     final effect = StreamChatTheme.of(context).colorTheme.borderBottom;
     return Container(
       height: 1,
+      // ignore: deprecated_member_use
       color: effect.color!.withOpacity(effect.alpha ?? 1.0),
     );
   }